% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{blarg}
\alias{blarg}
\alias{blarg2}
\alias{blarg3}
\alias{blarg4}
\title{Silly little wrapper to data.table}
\usage{
blarg(...)

blarg2(...)

blarg3(...)

blarg4(...)
}
\arguments{
\item{...}{passed along to data.table::data.table}
}
\description{
Silly little wrapper to data.table
}
