% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{extract}
\alias{extract}
\alias{[.blarg}
\alias{[.blarg4}
\alias{[.blarg2}
\title{Custom extract function (or subset or whatever) for dtsurveys. A loose wrapper over `[.data.table`}
\usage{
\method{[}{blarg}(x, i, j, by, ...)

\method{[}{blarg4}(x, i, j, by, ...)

\method{[}{blarg2}(x, i, j, by, ...)
}
\arguments{
\item{x}{a dtsurvey object}

\item{i}{i in the data.table format}

\item{j}{j in the data.table format}

\item{by}{by in the data.table format}

\item{...}{extra options passed to data table}
}
\description{
Custom extract function (or subset or whatever) for dtsurveys. A loose wrapper over `[.data.table`

Custom extract function (or subset or whatever) for dtsurveys. A loose wrapper over `[.data.table`
}
